Fix Icons for FeatureMarkers
This is just the start of a fix, with proof of concept showing different markes for different features.

Could store the icon names (paths) as a property in the database and return those to expedite the marker calls.  But this works for now and will require more IF statements to handle all cases.